pull out of utils / only used in aws provider
