+++ iterative/utils/helpers.go
@@ -7,17 +7,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-var (
-	AWSSynthRegions = map[string]map[string]string{
-		"aws": {
-			"us-east":  "us-east-1",
-			"us-west":  "us-west-1",
-			"eu-north": "eu-north-1",
-			"eu-west":  "eu-west-1",
-		},
-	}
-)
-
 func MachinePrefix(d *schema.ResourceData) string {
 	prefix := ""
 	if _, hasMachine := d.GetOk("machine"); hasMachine {
@@ -35,27 +24,6 @@ func SetId(d *schema.ResourceData) {
 			d.Set("name", d.Id())
 		}
 	}
-}
-
-func StripAvailabilityZone(region string) string {
-	lastChar := region[len(region)-1]
-	if lastChar >= 'a' && lastChar <= 'z' {
-		return region[:len(region)-1]
-	}
-	return region
-}
-
-func AWSGetRegion(d *schema.ResourceData) string {
-	region := d.Get("region").(string)
-	cloud := d.Get("cloud").(string)
-
-	if val, ok := AWSSynthRegions[cloud][region]; ok {
-		return val
-	}
-	if cloud == "aws" {
-		return StripAvailabilityZone(region)
-	}
-	return region
 }
 
 func LoadGCPCredentials() string {
